+++ pkg/resources/resource_scim_group_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"net/http"
 	"testing"
-	"time"
 
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/clients"
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/testhelpers"
@@ -25,9 +24,8 @@ func TestScimGroupResourceCreate(t *testing.T) {
 
 	testhelpers.WithMockFronteggServer(t, func(serverURL string) {
 		client := &clients.FronteggClient{
-			Endpoint:    serverURL,
-			HTTPClient:  &http.Client{},
-			TokenExpiry: time.Date(9999, 1, 1, 0, 0, 0, 0, time.UTC),
+			Endpoint:   serverURL,
+			HTTPClient: &http.Client{},
 		}
 
 		providerMeta := &utils.ProviderMeta{
@@ -50,9 +48,8 @@ func TestScimGroupResourceRead(t *testing.T) {
 
 	testhelpers.WithMockFronteggServer(t, func(serverURL string) {
 		client := &clients.FronteggClient{
-			Endpoint:    serverURL,
-			HTTPClient:  &http.Client{},
-			TokenExpiry: time.Date(9999, 1, 1, 0, 0, 0, 0, time.UTC),
+			Endpoint:   serverURL,
+			HTTPClient: &http.Client{},
 		}
 
 		providerMeta := &utils.ProviderMeta{
@@ -78,9 +75,8 @@ func TestScimGroupResourceDelete(t *testing.T) {
 
 	testhelpers.WithMockFronteggServer(t, func(serverURL string) {
 		client := &clients.FronteggClient{
-			Endpoint:    serverURL,
-			HTTPClient:  &http.Client{},
-			TokenExpiry: time.Date(9999, 1, 1, 0, 0, 0, 0, time.UTC),
+			Endpoint:   serverURL,
+			HTTPClient: &http.Client{},
 		}
 
 		providerMeta := &utils.ProviderMeta{