+++ pkg/resources/resource_scim_group_users_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"net/http"
 	"testing"
-	"time"
 
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/clients"
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/testhelpers"
@@ -25,9 +24,8 @@ func TestScimGroupUsersCreate(t *testing.T) {
 
 	testhelpers.WithMockFronteggServer(t, func(serverURL string) {
 		client := &clients.FronteggClient{
-			Endpoint:    serverURL,
-			HTTPClient:  &http.Client{},
-			TokenExpiry: time.Date(9999, 1, 1, 0, 0, 0, 0, time.UTC),
+			Endpoint:   serverURL,
+			HTTPClient: &http.Client{},
 		}
 
 		providerMeta := &utils.ProviderMeta{
@@ -48,9 +46,8 @@ func TestScimGroupUsersRead(t *testing.T) {
 
 	testhelpers.WithMockFronteggServer(t, func(serverURL string) {
 		client := &clients.FronteggClient{
-			Endpoint:    serverURL,
-			HTTPClient:  &http.Client{},
-			TokenExpiry: time.Date(9999, 1, 1, 0, 0, 0, 0, time.UTC),
+			Endpoint:   serverURL,
+			HTTPClient: &http.Client{},
 		}
 
 		providerMeta := &utils.ProviderMeta{
@@ -72,9 +69,8 @@ func TestScimGroupUsersUpdate(t *testing.T) {
 
 	testhelpers.WithMockFronteggServer(t, func(serverURL string) {
 		client := &clients.FronteggClient{
-			Endpoint:    serverURL,
-			HTTPClient:  &http.Client{},
-			TokenExpiry: time.Date(9999, 1, 1, 0, 0, 0, 0, time.UTC),
+			Endpoint:   serverURL,
+			HTTPClient: &http.Client{},
 		}
 
 		providerMeta := &utils.ProviderMeta{
@@ -98,9 +94,8 @@ func TestScimGroupUsersDelete(t *testing.T) {
 
 	testhelpers.WithMockFronteggServer(t, func(serverURL string) {
 		client := &clients.FronteggClient{
-			Endpoint:    serverURL,
-			HTTPClient:  &http.Client{},
-			TokenExpiry: time.Date(9999, 1, 1, 0, 0, 0, 0, time.UTC),
+			Endpoint:   serverURL,
+			HTTPClient: &http.Client{},
 		}
 
 		providerMeta := &utils.ProviderMeta{